Cleaned up existing OpenWebHelp functions
